backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openpyxl
reportlab
google-genai
python-multipart
pydantic
python-dotenv